+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.3.0] - version alpha:
+* add PasswordElement
+* make groupelement inside container
+* customise groupelement design
 ## [0.1.1] - version alpha:
  * format files
  * fix code errors